+++ src/main/java/um/tesoreria/core/hexagonal/ubicacion/domain/ports/in/GetAllUbicacionesUseCase.java
@@ -1,0 +1,8 @@
+package um.tesoreria.core.hexagonal.ubicacion.domain.ports.in;
+
+import um.tesoreria.core.hexagonal.ubicacion.domain.model.Ubicacion;
+import java.util.List;
+
+public interface GetAllUbicacionesUseCase {
+    List<Ubicacion> getAllUbicaciones();
+}