+++ src/main/java/um/tesoreria/core/hexagonal/ubicacion/domain/ports/out/UbicacionRepository.java
@@ -1,0 +1,9 @@
+package um.tesoreria.core.hexagonal.ubicacion.domain.ports.out;
+
+import um.tesoreria.core.hexagonal.ubicacion.domain.model.Ubicacion;
+import java.util.List;
+
+public interface UbicacionRepository {
+    List<Ubicacion> findAll();
+    List<Ubicacion> findAllByGeograficaId(Integer geograficaId);
+}