+++ src/main/java/um/tesoreria/core/hexagonal/ubicacionArticulo/infrastructure/web/dto/UbicacionArticuloRequest.java
@@ -1,0 +1,11 @@
+package um.tesoreria.core.hexagonal.ubicacionArticulo.infrastructure.web.dto;
+
+import lombok.Data;
+import java.math.BigDecimal;
+
+@Data
+public class UbicacionArticuloRequest {
+    private Integer ubicacionId;
+    private Long articuloId;
+    private BigDecimal numeroCuenta;
+}