+++ src/main/java/um/tesoreria/core/model/dto/AsignacionCostoDto.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 
 import um.tesoreria.core.hexagonal.articulo.infrastructure.persistence.entity.ArticuloEntity;
 import um.tesoreria.core.kotlin.model.Comprobante;
-import um.tesoreria.core.model.UbicacionArticulo;
+import um.tesoreria.core.hexagonal.ubicacionArticulo.infrastructure.persistence.entity.UbicacionArticuloEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -33,7 +33,7 @@ public class AsignacionCostoDto implements Serializable {
 	private ProveedorArticulo proveedorArticulo;
 	private ArticuloEntity articulo;
 	private Comprobante comprobante;
-	private UbicacionArticulo ubicacionArticulo;
+	private UbicacionArticuloEntity ubicacionArticulo;
 
 	private BigDecimal importe;